IoT/testtest/requirements.txt
Django==2.1.15
django-cors-headers==3.4.0
pyserial==3.4
pytz==2020.1
